move rewriter logic
