+++ packages/proxy/lib/http/util/service-worker.ts
@@ -1,5 +1,3 @@
-/// <reference lib="dom" />
-
 import Debug from 'debug'
 import pDefer from 'p-defer'
 import type { BrowserPreRequest } from '../../types'
@@ -57,113 +55,6 @@ export const serviceWorkerFetchEventHandler = (handler: ServiceWorkerFetchHandle
       handler({ url, isControlled: respondWithCalled })
     }
   }
-}
-
-/**
- * Rewrites the service worker to listen to fetch events to determine if the service worker handled the request.
- * @param body the body of the service worker to rewrite
- * @returns the rewritten service worker
- */
-export const rewriteServiceWorker = (body: Buffer) => {
-  function __cypressCreateListenerFunction (listener) {
-    return (event) => {
-      // we want to override the respondWith method so we can track if it was called
-      // to determine if the service worker handled the request
-      const oldRespondWith = event.respondWith
-      let respondWithCalled = false
-
-      event.respondWith = (...args) => {
-        respondWithCalled = true
-        oldRespondWith.call(event, ...args)
-      }
-
-      const returnValue = listener(event)
-
-      // @ts-expect-error
-      // call the CDP binding to inform the backend whether or not the service worker handled the request
-      self.__cypressServiceWorkerFetchEvent(JSON.stringify({ url: event.request.url, respondWithCalled }))
-
-      return returnValue
-    }
-  }
-
-  function __cypressOverwriteAddRemoveEventListener () {
-    const listeners = new WeakMap()
-
-    const oldAddEventListener = self.addEventListener
-
-    // Overwrite the addEventListener method so we can
-    // determine if the service worker handled the request
-    self.addEventListener = (type, listener, options) => {
-      if (type === 'fetch') {
-        const newListener = __cypressCreateListenerFunction(listener)
-
-        listeners.set(listener, newListener)
-
-        return oldAddEventListener(type, newListener, options)
-      }
-
-      return oldAddEventListener(type, listener, options)
-    }
-
-    const oldRemoveEventListener = self.removeEventListener
-
-    // Overwrite the removeEventListener method so we can
-    // remove the listener from the map
-    self.removeEventListener = (type, listener, options) => {
-      if (type === 'fetch') {
-        const newListener = listeners.get(listener)
-
-        listeners.delete(listener)
-
-        return oldRemoveEventListener(type, newListener, options)
-      }
-
-      return oldRemoveEventListener(type, listener, options)
-    }
-  }
-
-  function __cypressOverwriteOnfetch () {
-    const originalPropertyDescriptor = Object.getOwnPropertyDescriptor(
-      self,
-      'onfetch',
-    )
-
-    if (!originalPropertyDescriptor) {
-      return
-    }
-
-    // Overwrite the onfetch property so we can
-    // determine if the service worker handled the request
-    Object.defineProperty(
-      self,
-      'onfetch',
-      {
-        configurable: originalPropertyDescriptor.configurable,
-        enumerable: originalPropertyDescriptor.enumerable,
-        get () {
-          return originalPropertyDescriptor.get?.call(this)
-        },
-        set (value: (event) => void) {
-          let newHandler
-
-          if (value) {
-            newHandler = __cypressCreateListenerFunction(value)
-          }
-
-          originalPropertyDescriptor.set?.call(this, newHandler)
-        },
-      },
-    )
-  }
-
-  const updatedBody = `
-${__cypressCreateListenerFunction};
-(${__cypressOverwriteAddRemoveEventListener})();
-(${__cypressOverwriteOnfetch})();
-${body}`
-
-  return updatedBody
 }
 
 /**
@@ -288,7 +179,7 @@ export class ServiceWorkerManager {
         activatedServiceWorker.scriptURL === paramlessDocumentURL ||
         !activatedServiceWorker.initiatorOrigin ||
         !paramlessDocumentURL.startsWith(activatedServiceWorker.initiatorOrigin)) {
-        debug('Service worker not activated or request URL is from the service worker, skipping: %o', { activatedServiceWorker, paramlessDocumentURL })
+        debug('Service worker not activated or request URL is from the service worker, skipping: %o', { activatedServiceWorker, browserPreRequest })
 
         return
       }
